+++ NEWS.md
@@ -25,6 +25,7 @@
 ## new
 
 - `animate<SpatVector>` method by Márcia Barbosa [#2007](https://github.com/rspatial/terra/issues/2007)
+- `agitate<SpatVector>` method to jitter a SpatVector of points
 
 
 # version 1.8-93